up: annotation - move Resource to Concern/AbstractResource
